+++ packages/cli/src/types/index.ts
@@ -44,6 +44,7 @@ export type TranslateFlags = {
   dryRun: boolean;
   saveLocal?: boolean;
   stageTranslations?: boolean;
+  setupProject?: boolean; // if true, skip enqueue step
   publish?: boolean;
   force?: boolean;
   forceDownload?: boolean;
@@ -268,6 +269,10 @@ export type SourceObjectOptions = {
   // optional config for transforming specific fields in the sourceItem
   // for example, helpful for handling urls with locale-specific paths
   transform?: TransformOptions;
+
+  // optional sorting behavior for array sourceObjects. When set to 'locale',
+  // the array will be ordered to match the locales array provided to mergeJson
+  experimentalSort?: 'locales';
 };
 
 export type TransformOptions = {