+++ packages/cli/src/workflow/PollJobsStep.ts
@@ -1,10 +1,9 @@
 import chalk from 'chalk';
 import { WorkflowStep } from './Workflow.js';
-import { createProgressBar, logError } from '../console/logging.js';
-import { getLocaleProperties } from 'generaltranslation';
+import { logger } from '../console/logger.js';
 import { GT } from 'generaltranslation';
 import { EnqueueFilesResult } from 'generaltranslation/types';
-import { TEMPLATE_FILE_NAME } from '../cli/commands/stage.js';
+import { TEMPLATE_FILE_NAME } from '../utils/constants.js';
 import type { FileProperties } from '../types/files.js';
 
 export type PollJobsInput = {
@@ -31,7 +30,7 @@ export class PollTranslationJobsStep extends WorkflowStep<
   PollJobsInput,
   PollJobsOutput
 > {
-  private spinner: ReturnType<typeof createProgressBar> | null = null;
+  private spinner: ReturnType<typeof logger.createProgressBar> | null = null;
   private previousProgress = 0;
 
   constructor(private gt: GT) {
@@ -46,7 +45,7 @@ export class PollTranslationJobsStep extends WorkflowStep<
     forceRetranslation,
   }: PollJobsInput): Promise<PollJobsOutput> {
     const startTime = Date.now();
-    this.spinner = createProgressBar(fileQueryData.length);
+    this.spinner = logger.createProgressBar(fileQueryData.length);
     const spinnerMessage = forceRetranslation
       ? 'Waiting for retranslation...'
       : 'Waiting for translation...';
@@ -212,7 +211,7 @@ export class PollTranslationJobsStep extends WorkflowStep<
               }
             }
           } catch (error) {
-            logError(chalk.red('Error checking job status: ') + error);
+            logger.error(chalk.red('Error checking job status: ') + error);
           }
         }, 5000);
       }, msUntilNextInterval);