+++ packages/next/CHANGELOG.md
@@ -1,5 +1,15 @@
 # gt-next
 
+## 6.9.2
+
+### Patch Changes
+
+- Updated dependencies [[`afbd29a`](https://github.com/generaltranslation/gt/commit/afbd29a34b051c76fce387269c4eb4a2e00a5831)]:
+  - generaltranslation@8.0.3
+  - @generaltranslation/compiler@1.1.4
+  - gt-react@10.8.2
+  - @generaltranslation/supported-locales@2.0.26
+
 ## 6.9.1
 
 ### Patch Changes